불필요 파일 삭제, 로그아웃 기능 추가

+++ src/main/java/com/pshc/util/dto/CategoryRepository.java
@@ -8,6 +8,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.pshc.util.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Long>{
-	
+public interface CategoryRepository extends JpaRepository<Category, Integer>{
 }
